Projects/06-Chat_With_Textbook/requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
Pillow>=10.0.0
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0 